+++ RadixWallet/Features/AccountPreferencesFeature/Children/DeleteAccount/ChooseReceivingAccountOnDelete+View.swift
@@ -84,15 +84,10 @@ private extension StoreOf<ChooseReceivingAccountOnDelete> {
 private extension View {
 	func destinations(with store: StoreOf<ChooseReceivingAccountOnDelete>) -> some View {
 		let destinationStore = store.destination
-		return confirmDeletionAlert(with: destinationStore)
-			.tooManyAssetsAlert(with: destinationStore)
+		return confirmSkipAlert(with: destinationStore)
 	}
 
-	private func confirmDeletionAlert(with destinationStore: PresentationStoreOf<ChooseReceivingAccountOnDelete.Destination>) -> some View {
+	private func confirmSkipAlert(with destinationStore: PresentationStoreOf<ChooseReceivingAccountOnDelete.Destination>) -> some View {
 		alert(store: destinationStore.scope(state: \.confirmSkipAlert, action: \.confirmSkipAlert))
-	}
-
-	private func tooManyAssetsAlert(with destinationStore: PresentationStoreOf<ChooseReceivingAccountOnDelete.Destination>) -> some View {
-		alert(store: destinationStore.scope(state: \.tooManyAssetsAlert, action: \.tooManyAssetsAlert))
 	}
 }
