+++ RadixWallet/Features/AccountPreferencesFeature/Children/DeleteAccount/DeleteAccountConfirmation+View.swift
@@ -42,30 +42,6 @@ extension DeleteAccountConfirmation {
 					}
 					.controlState(viewStore.footerButtonState)
 				}
-				.destination(store: store)
-			}
-		}
-	}
-}
-
-@MainActor
-private extension View {
-	func destination(store: StoreOf<DeleteAccountConfirmation>) -> some View {
-		let destinationStore = store.scope(state: \.$destination, action: \.destination)
-		return chooseReceivingAccount(with: destinationStore, store: store)
-			.accountDeleted(with: destinationStore, store: store)
-	}
-
-	private func chooseReceivingAccount(with destinationStore: PresentationStoreOf<DeleteAccountConfirmation.Destination>, store: StoreOf<DeleteAccountConfirmation>) -> some View {
-		navigationDestination(store: destinationStore.scope(state: \.chooseReceivingAccount, action: \.chooseReceivingAccount)) {
-			ChooseReceivingAccountOnDelete.View(store: $0)
-		}
-	}
-
-	private func accountDeleted(with destinationStore: PresentationStoreOf<DeleteAccountConfirmation.Destination>, store: StoreOf<DeleteAccountConfirmation>) -> some View {
-		fullScreenCover(store: destinationStore.scope(state: \.accountDeleted, action: \.accountDeleted)) { _ in
-			AccountDeletedView {
-				store.send(.view(.goHomeButtonTapped))
 			}
 		}
 	}
